+++ bindings/python/multibody/expose-frame.cpp
@@ -7,7 +7,7 @@
 #include "pinocchio/bindings/python/fwd.hpp"
 #include "pinocchio/bindings/python/multibody/frame.hpp"
 #include "pinocchio/bindings/python/serialization/serialization.hpp"
-#include "pinocchio/bindings/python/utils/std-aligned-vector.hpp"
+#include "pinocchio/bindings/python/utils/std-vector.hpp"
 
 namespace pinocchio
 {
@@ -17,9 +17,9 @@ namespace pinocchio
     void exposeFrame()
     {
       FramePythonVisitor<context::Frame>::expose();
-      StdAlignedVectorPythonVisitor<context::Frame>::expose("StdVec_Frame");
+      StdVectorPythonVisitor<std::vector<context::Frame>>::expose("StdVec_Frame");
 #ifndef PINOCCHIO_PYTHON_NO_SERIALIZATION
-      serialize<StdAlignedVectorPythonVisitor<context::Frame>::vector_type>();
+      serialize<std::vector<context::Frame>>();
 #endif
     }