+++ bindings/python/spatial/expose-force.cpp
@@ -7,7 +7,7 @@
 #include "pinocchio/bindings/python/fwd.hpp"
 #include "pinocchio/bindings/python/spatial/force.hpp"
 #include "pinocchio/bindings/python/serialization/serialization.hpp"
-#include "pinocchio/bindings/python/utils/std-aligned-vector.hpp"
+#include "pinocchio/bindings/python/utils/std-vector.hpp"
 
 namespace pinocchio
 {
@@ -17,9 +17,9 @@ namespace pinocchio
     void exposeForce()
     {
       ForcePythonVisitor<context::Force>::expose();
-      StdAlignedVectorPythonVisitor<context::Force>::expose("StdVec_Force");
+      StdVectorPythonVisitor<std::vector<context::Force>>::expose("StdVec_Force");
 #ifndef PINOCCHIO_PYTHON_NO_SERIALIZATION
-      serialize<StdAlignedVectorPythonVisitor<context::Force>::vector_type>();
+      serialize<std::vector<context::Force>>();
 #endif
     }