+++ bindings/python/spatial/expose-symmetric3.cpp
@@ -7,7 +7,7 @@
 #include "pinocchio/bindings/python/fwd.hpp"
 #include "pinocchio/bindings/python/spatial/symmetric3.hpp"
 #include "pinocchio/bindings/python/serialization/serialization.hpp"
-#include "pinocchio/bindings/python/utils/std-aligned-vector.hpp"
+#include "pinocchio/bindings/python/utils/std-vector.hpp"
 
 namespace pinocchio
 {
@@ -17,9 +17,9 @@ namespace pinocchio
     void exposeSymmetric3()
     {
       Symmetric3PythonVisitor<context::Symmetric3>::expose();
-      StdAlignedVectorPythonVisitor<context::Symmetric3>::expose("StdVec_Symmetric3");
+      StdVectorPythonVisitor<std::vector<context::Symmetric3>>::expose("StdVec_Symmetric3");
 #ifndef PINOCCHIO_PYTHON_NO_SERIALIZATION
-      serialize<StdAlignedVectorPythonVisitor<context::Symmetric3>::vector_type>();
+      serialize<std::vector<context::Symmetric3>>();
 #endif // ifndef PINOCCHIO_PYTHON_NO_SERIALIZATION
     }