+++ CHANGELOG.md
@@ -57,6 +57,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Remove support to eigenpy < 3
 - Don't use Eigen::aligned_allocator:
   - Remove header `pinocchio/serialization/aligned-vector.hpp`
+  - Remove header `pinocchio/bindings/python/utils/std-aligned-vector.hpp`
 
 ## [3.9.0] - 2026-01-05